IOSS: move compare methods to Ioss_Compare.[hC] IOSS: add Ioss_Compare.h to Ioss_SubSystem.h
