+++ packages/seacas/libraries/ioss/src/Ioss_Utils.h
@@ -501,11 +501,6 @@ namespace Ioss {
                               const std::string &header, const std::string &suffix = "\n\t",
                               bool print_empty = false);
 
-    // Compare the mesh in 'input_region' to 'output_region'.  Behavior can be controlled
-    // via options in 'options'
-    static bool compare_database(Ioss::Region &input_region, Ioss::Region &output_region,
-                                 const Ioss::MeshCopyOptions &options);
-
   };
 
   inline std::ostream &OUTPUT() { return *Utils::m_outputStream; }
